Extract jsx element generation  from presetnter
