+++ src/Web/ClientApp/src/components/pages/InspectionTypeCategory.tsx
@@ -1,18 +1,19 @@
 import React, { FC, useState, useEffect } from "react";
-import { Grid, Paper, TableContainer } from "@mui/material";
+import { Grid, Paper, Table, TableBody, TableCell, TableContainer, TableHead, TableRow } from "@mui/material";
 import nameof from "ts-nameof.macro";
 import {
   Notification,
   NotificationInitState,
   NotificationStateInteractor,
 } from "../utilities/Notification";
-import { BottomNavigationAdd, TopPageLink } from "../utilities";
+import { BottomNavigationAdd, CancelIconButton, EditIconButton, TopPageLink } from "../utilities";
 import { EditDialog } from "../dialog/EditDialog";
 import {
   IInspectionTypeController,
   IInspectionTypePresenter,
 } from "../../interfaces";
 import { useDIContext } from "../../container";
+import { InspectionType } from "../../entities";
 
 export const InspectionTypeCategory: FC = (): JSX.Element => {
   const inject = useDIContext();
@@ -104,7 +105,32 @@ export const InspectionTypeCategory: FC = (): JSX.Element => {
         </Grid>
         <Grid item xs={12}>
           <TableContainer component={Paper}>
-            {presenter.inspectionTypeTable(handleUpdateItem, handleDeleteItem)}
+          <Table>
+        <TableHead>
+          <TableRow>
+            <TableCell>点検タイプ</TableCell>
+            <TableCell>&nbsp;</TableCell>
+            <TableCell>&nbsp;</TableCell>
+          </TableRow>
+        </TableHead>
+        <TableBody>
+          {presenter.state.map((type: InspectionType) => (
+            <TableRow key={type.inspectionTypeId}>
+              <TableCell>{type.description}</TableCell>
+              <TableCell padding="checkbox">
+                <EditIconButton
+                  onClick={() => handleUpdateItem(type.inspectionTypeId)}
+                />
+              </TableCell>
+              <TableCell padding="checkbox">
+                <CancelIconButton
+                  onClick={() => handleDeleteItem(type.inspectionTypeId)}
+                />
+              </TableCell>
+            </TableRow>
+          ))}
+        </TableBody>
+      </Table>
           </TableContainer>
         </Grid>
         <Grid item xs={12}>
