+++ src/Web/ClientApp/src/presenters/InspectionGroupPresenter.tsx
@@ -1,13 +1,5 @@
-import {
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TableRow,
-} from "@mui/material";
 import { IInspectionGroupInteractor } from "../interfaces";
 import { InspectionGroup } from "../entities";
-import { CancelIconButton, EditIconButton } from "../components/utilities";
 import { IInspectionGroupPresenter } from "../interfaces/presenter";
 
 export class InspectionGroupPresenter implements IInspectionGroupPresenter {
@@ -27,43 +19,5 @@ export class InspectionGroupPresenter implements IInspectionGroupPresenter {
     return this.useCase.groups.find(
       (x: InspectionGroup) => x.inspectionGroupId === id
     )?.description;
-  }
-
-  inspectionGroupTable(
-    updateMethod: (id: number) => void,
-    deleteMethod: (id: number) => void
-  ): JSX.Element {
-    return (
-      <Table>
-        <TableHead>
-          <TableRow>
-            <TableCell>ID</TableCell>
-            <TableCell>点検グループ</TableCell>
-            <TableCell>&nbsp;</TableCell>
-            <TableCell>&nbsp;</TableCell>
-          </TableRow>
-        </TableHead>
-        <TableBody>
-          {this.useCase.groups.map((type: InspectionGroup) => (
-            <TableRow key={type.inspectionGroupId}>
-              <TableCell padding="checkbox" align="center">
-                {type.inspectionGroupId}
-              </TableCell>
-              <TableCell>{type.description}</TableCell>
-              <TableCell padding="checkbox">
-                <EditIconButton
-                  onClick={() => updateMethod(type.inspectionGroupId)}
-                />
-              </TableCell>
-              <TableCell padding="checkbox">
-                <CancelIconButton
-                  onClick={() => deleteMethod(type.inspectionGroupId)}
-                />
-              </TableCell>
-            </TableRow>
-          ))}
-        </TableBody>
-      </Table>
-    );
   }
 }
